Move portfolio_input_check() to file with its name
This commit renames a file to match the name of a main
function, and move that main function to such file.

It's common practice for a main function to live in a file
with a matching name. For reference see the guidelines below,
which donn't exactly match this case but close enough.

> * If a file contains a single function, give the file the
    same name as the function.
> * If a file contains multiple related functions, give it a
     concise, but evocative name.
>
> -- https://style.tidyverse.org/package-files.html