Move R/script.R to vignettes/
In a package, code that is meant to run as a script can be shown
as a vignette. Here I'm using a flavour of a vignette called
"article" that shows up only on the package website but not as an
installed file in the package. This saves space and makes the
package installation faster.

An alternative is to show that code in README.Rmd, which makes it
the home page of the package website. This is appropriate if such
code is a good example of what the package does.