Replace first_chr_up() with str_to_sentence()
Here I am reusing the robust package stringr to minimize the
chances of introducing bugs. This comes at the cost of depending
on a package that we were not useing before (strignr), but whether
this is a problem or not is not obvious -- it depends
(https://www.tidyverse.org/blog/2019/05/itdepends/).

```
stringr::str_to_sentence(
  c("ab cd", "ef gh")
)
> [1] "Ab cd" "Ef gh"
```