moved brown_technologies_list outside and made it a function input
The high-carbon-list exists already in the parameter file and will be used at several places within the analysis. Thus it should be an input to any function to ensure consistency across different analyses.
Also the list might change and we do not want to change the functions everytime this happens.