Polish help file
* Delegate clarification of data arguments

This is intentionally vague, and delegates further details to the
documentation of the `market` and `portfolio` datasets. This is so
I can get this helpfile close to ready even when I have missing
information, i.e. What does it mean for a dataframe to be of the
kind "market results" or "portfolio results"?

@vintented, please either edit the description here so it stands
alone or add a more complete description in the documentation of
`market` and `portfolio`.

Also:

* Explain deafaults.

* Fix reference to other docs

* Update title and descritioin
  Follows edit by @2diiKlaus via
  https://github.com/2DegreesInvesting/r2dii.analysis/
  pull/8#issuecomment-562592954