Replace .funs = funs(.) by .funs = ~ lambda_fun(.)
test() shows:

test-single_indicator.R:22: warning: calculate_temperature_indicator and
friends yield known values
funs() is soft deprecated as of dplyr 0.8.0
Please use a list of either functions or lambdas ...

I search for "funs(" and I see what seems to be an incorrect use
of funs().

Instead the `.funs` argument to `rename_at` expects a lambda with
the form ~ fun(.) such as:

    .funs = ~ paste0("a", .x)

or a normal function such as:

    .funs = function(.x) paste0("a", .x)

test() no longer throws this warning, and check() no longer notices
`funs` as an undefined global variable.