Use defaults more directly to clarify signature
I think this is a bit more readable. The %||% operator is mostly
to protect the users from implementation details, but in tests/ the
users are the developers themselves, so I think in this case it's
prefferable to expose the implementation details so the reader of
the test can see what's going on at a glance.