Use `interesting_scenario_columns()` if present
`one_of()` allows us to pick columns that are present and pass
if any column is not present. But warns about the columns that are
not present.

That warning seems needless and problematic:
* Needless because I see no prolem in returning what the function
  gets, and not returning what the function never got.
* Also needless because crucial columns are already checked.
* Problematic because we need to suppress the warning in every
  test that uses fake data -- which doesn't have all the
  columns in `interesting_scenario_columns()`.

I conclude that the warning should suppressed, or lowered to a
message (which isn't problematic in tests).