Rename to join_ald_scenario()
The signature of join_portfolio_ald_scenario() has no reference
to  "porftolio".

Sometimes the code is clearer if the name of a function refers to
the second (or subsequent) arguments rather than the first. The
first argument is usually clear enough from the name of the object
fed to it, or because it is obviously related to what the goal of
the function.

Example:

data %>% sample_n(size = 10)
\# not `data %>% dplyr::sample_n_from_data(size = 10)`

data %>% tibble::add_column(new = "some value")
\# not `data %>% tibble::add_column_to_data(new "some value")`

In this case:
validated_match %>% join_ald_scenario(ald, scenario)