Focus on usability until we have evidence for speed
Until we have some evidence that speed is a problem, I suggest we
focus on how easy the code is to understand and run. Here I use the
pipie to emphasize the step-like workflow.

@jdhoffa, you said that some ald datasets are big, but the data we
would feed to add_fare_share_columns() would be first reduced in
size by match_name(), then prioritize(), right? So the data may
likely be small enougth that we should not worry. (If the data is
large, then an even bigger problem would be that users take too
long to validate the data.) Do I understand this well?