Avoid clash in column name 'source'
This commit tracks the provenance of the column "source" back to
the dataset region_isos. This avoids a names clash with the column
"source" from the data that join_ald_scenario() modifies.

This is not just cosmetic but a fix. We recently modified
`r2dii.data::region_isos` and that caused a test to break. This
commit fixes that problem.