Style article
Readers expect the topic of the document to be on
the last sentence of the firts paragraph. In all other
paragraphs (except maybe the last) readers expect the
topic to be on the last sentence of the issue -- which
is sometimes the first sentence of the paragraph, in
which case the issue and the topic sentence are the same
sentence.