Refactor
This refactoring aims to find a simple way to express which
columns to unnest. The solution with where_() is quite elegant,
but we need to ensure that the only list columns in the dataset
are the ones we want to unnest -- hence we must assert that the
data has no list columns before calling pivot_wider().

where() from the tidyverse seems to be unstable -- but it's simple
so I reproduce it here under the name where_().