target_market_share() may lack some scenario (#204)
Closes #203.

`target_market_share()` now works as expected when some value
of the column `scenario` is missing for some value of the column
`region`. It no longer results in output columns `production` and
`technology_share` of type "list".

The refactoring aims to find a simple way to express which
columns to `unnest()`. The solution with `where_()` is quite elegant,
but we need to ensure that the only list columns in the dataset
are the ones we want to `unnest()` -- hence we must assert that the
data has no list columns before calling `pivot_wider()`.

`where()` from the tidyverse seems to be unstable (I struggled to find
an exported version); but it's simple to reproduce, so I copy-pasted it
from `tidyselect:::where()` and named it `where_()`.

Co-authored-by: Mauro Lepore <maurolepore@gmail.com>