Namespace pull() to enable `test_file()` (#289)
When running the entire test suite, the namespace of `pull()` is available (likely via DESCRIPTION). But this isn't the case when running `test_file()` (usually via shift + control + T), which results in a failing test (false positive). This commit removes that friction. An alternative if to use `library(dplyr)` at the top of the file -- but I'm now using the more "focal" approach.