thank in situ (#301)
* Thanks in situ

The acknowledgements were stored in a central repository and
pulled from all r2dii packages. This is DRY but problematic
because URLs are sometimes redirected, and the fix via urlchecker
happens in the downstream package, not in the upstream source of
the acknowledgement. Given that this caused a rejection on CRAN,
I prefer to keep the message WET in favor of local maintainability.

* Update readme