Increase in green technologies for pure coalcap player
I have changed the way scenario and region are joined to the ald
database. First I joined by region, then for each region I look for
each company which green technologies are missing and add them (with 0
as production) for each year of the analysis. The loop with automotive
technology is to avoid to add some lines for electric_hdv and hybrid_hdv
for a pure ldv player. Then I join scenario the same way it was done
before, but with two more columns to be matched (scenario source and
regions). This code can clearly be improved as it is not really clean
and also based on the ald_demo dataset that overcome 2025 while it is
usually not the case (#240 )