ADO Bug 876 - NAs in technology share
Add some conditions where techmix was NA to be 0. In my understanding of techmix it doesn't change the techmix value to add some zero in the formula. I would feel more confortable about the calculation of this metric if we did the pivot wider before this then filter by tech_mix=0 but it request far more work.
The lines might be more efficient but I become mad as I didn't arrive to do it in one line.